+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v1.1.3 - 2022/11/28
+
+* [Added] Added the submenu open state.
+  * Added the css `.mx-context-menu-item.open` indicates that the submenu of the current menu is open.
+  * Added the property `isOpen` in render slot data (`MenuItemRenderData`) indicates that the submenu of the current menu is open.
+  * Added the property `hasChildren` in render slot data (`MenuItemRenderData`) indicates whether the current menu has submenus.
+* [Added] Added the svg icon support. (`svgIcon` and `svgProps` property).
+* [Change] The parent menu will now remain highlighted when the submenu open.
+
 ## v1.1.2 - 2022/11/06
 
 * Fix the problem that the binding event was not removed correctly. (#14 Thanks @liumingye)